backend/app/services/crm_service.py
import os
from typing import Dict, Any
from dotenv import load_dotenv

load_dotenv()

# Placeholder for CRM API Key or authentication details
CRM_API_KEY = os.getenv("CRM_API_KEY")
CRM_API_BASE_URL = os.getenv("CRM_API_BASE_URL")

class CRMService:
    def __init__(self):
        if not CRM_API_KEY or not CRM_API_BASE_URL:
            print("Warning: CRM_API_KEY or CRM_API_BASE_URL not set. CRM integration will use dummy data.")

    async def send_to_crm(self, crm_data: Dict[str, Any], entity_type: str) -> Dict[str, Any]:
        """
        Sends data to the CRM.
        This is a placeholder function and would be replaced with actual CRM API calls.
        """
        print(f"Simulating sending {entity_type} data to CRM: {crm_data}")
        # In a real application, you would make an HTTP request to your CRM API here.
        # Example: response = await httpx.post(f"{CRM_API_BASE_URL}/api/{entity_type}", json=crm_data, headers={"Authorization": f"Bearer {CRM_API_KEY}"})
        # response.raise_for_status()
        # return response.json()
        
        # For now, just return the data as if it was successfully sent
        return {"status": "success", "message": f"{entity_type} sent to CRM successfully (simulated)", "data": crm_data} 