backend/requirements.txt
fastapi
uvicorn
pymongo==4.7.3
motor==3.3.2
email-validator==2.1.1
python-dotenv==1.0.1
